Address PR comments: move history-util to datasets/
